renomear cena 2 para cena 3 Fixes #164
