+++ src/config.ts
@@ -1,11 +1,9 @@
 import { Types } from "phaser";
 import { Scene0 } from "./scenes/scene0";
 import { Scene1 } from "./scenes/scene1";
-import { Scene2 } from "./scenes/scene2";
 import { Scene3 } from "./scenes/scene3";
 import { Scene4 } from "./scenes/scene4";
-import { Scene5 } from "./scenes/scene56";
-import { Scene6 } from "./scenes/scene6";
+import { Scene56 } from "./scenes/scene56";
 import { Scene7 } from "./scenes/scene7";
 import { Scene8 } from "./scenes/scene8";
 
@@ -20,11 +18,9 @@ const config: Types.Core.GameConfig = {
   scene: [
     Scene0,
     Scene1,
-    Scene2,
     Scene3,
     Scene4,
-    Scene5,
-    Scene6,
+    Scene56,
     Scene7,
     Scene8,
   ],
