+++ static/assets/scene3/clear.png
